requirements.txt
selenium
webdriver-manager
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
oauth2client
gspread-dataframe
openpyxl
xlrd>=2.0.1
